pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qudata"
version = "1.0.2"
description = "Data processing pipeline for LLM training datasets"
authors = [{name = "Qub△se Team"}]
license = "MIT"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "pdfplumber>=0.10.0",
    "python-docx>=0.8.11",
    "nltk>=3.8.0",
    "spacy>=3.6.0",
    "langdetect>=1.0.9",
    "chardet>=5.0.0",
    "pytesseract>=0.3.10",
    "opencv-python>=4.8.0",
    "Pillow>=10.0.0",
    "feedparser>=6.0.10",
    "kafka-python>=2.0.2",
    "plotly>=5.15.0",
    "streamlit>=1.25.0",
    "jinja2>=3.1.0",
    "reportlab>=4.0.0",
    "croniter>=1.4.0",
    "PyMuPDF>=1.24.0",
    # API server & system metrics
    "fastapi>=0.110.0",
    "psutil>=5.9.0",
    "uvicorn[standard]>=0.23.0",
    "gunicorn>=21.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

web = [
    "scrapy>=2.10.0",
    "selenium>=4.10.0",
]

ml = [
    "scikit-learn>=1.3.0",
    "transformers>=4.30.0",
    "bertopic>=0.15.0",
]

[project.scripts]
qudata = "qudata.cli:main"

[project.urls]
Homepage = "https://github.com/qubasehq/qudata"
Repository = "https://github.com/qubasehq/qudata"
Issues = "https://github.com/qubasehq/qudata/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
